requirements.txt
django
mysqlclient
django-crispy-forms
pillow
django-admin-interface
django-widget-tweaks
django-localflavor
localflavor
reportlab
djangorestframework
